Migrate the test
