+++ tests/wasm/main.rs
@@ -6,3 +6,4 @@ extern crate wasm_bindgen;
 
 pub mod api;
 pub mod option;
+pub mod optional_numbers;
